Error Handling
Added error checking for accessing an exam. Makes sure that the current time is within the start and close time of the exam, the question is in the exam, the student is in the section the exam is for,  as well as error handling for if anything comes back null. Also added a way to get indivual start and close time for a specific user on a specific exam.