Error Handling/Grades Stuff
Changed permission of grade routes so any logged in user can use the route. Added checks for accessing grades so only a ta or teacher can see grades other than their own. A ta now as the manage option when they are a ta of a section. The only option they will have is to see the grades of the section that they are a ta of.