Tracking Leave Page Occurrences
Now logs an occurence for a user, question, and exam with the current time everytime the window loses focus while in an exam. Added controller=question action=read_occurences and action=write_occurence. Put a bunch of unsets in question_controller to help not have so many variables in memory at a time.  Updated route permissions. Created view to display occurences. Put 'occurences' under manage tab for teachers and ta's.